services/ml-inference/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
numpy==2.2.2
tensorflow>=2.15.0
